Remove in-memory cache for storing JWT signing key
The JWT signing key is derived from `JWT_SECRET` and passed to our code
by next-auth in the `encode` and `decode` callbacks.

In order to make the secret available to other callbacks, I was wrongly
storing it in a memory cache, which simply doesn't work in serveless
deployments like vercel.

Switch to using redis as the cache.